test: Consolidating Part tests.
